captcha_test/requirements.txt
onnxruntime>=1.15.0
torchvision>=0.15.0
pillow>=10.0.0
requests>=2.31.0
pycryptodome>=3.19.0
